코드 및 파일 구조 간소화
불필요한 요소 삭제, 구조 재구성
